+++ go.mod
@@ -194,4 +194,4 @@ replace google.golang.org/grpc v1.51.0 => google.golang.org/grpc v1.26.0
 // After the PR to kvproto is merged, remember to comment this out and run `go mod tidy`.
 // replace github.com/pingcap/kvproto => github.com/$YourPrivateRepo $YourPrivateBranch
 
-replace github.com/pingcap/kvproto => github.com/rleungx/kvproto v0.0.0-20230223090544-2d5287a102d2
+replace github.com/pingcap/kvproto => github.com/rleungx/kvproto v0.0.0-20230224062044-754da3291461
