+++ server/grpc_service.go
@@ -2007,118 +2007,6 @@ func (s *GrpcServer) WatchGlobalConfig(req *pdpb.WatchGlobalConfigRequest, serve
 	}
 }
 
-// Watch watches the key with a given prefix and revision.
-func (s *GrpcServer) Watch(req *pdpb.WatchRequest, server pdpb.PD_WatchServer) error {
-	ctx, cancel := context.WithCancel(s.Context())
-	defer cancel()
-	key := string(req.GetKey())
-	endKey := string(req.GetRangeEnd())
-	startRevision := req.GetStartRevision()
-	watchChan := s.client.Watch(ctx, key, clientv3.WithPrefix(), clientv3.WithRange(endKey), clientv3.WithRev(startRevision), clientv3.WithPrevKV())
-	for {
-		select {
-		case <-ctx.Done():
-			return nil
-		case res := <-watchChan:
-			if res.Err() != nil {
-				var resp pdpb.WatchResponse
-				if startRevision < res.CompactRevision {
-					resp.Header = s.wrapErrorToHeader(pdpb.ErrorType_DATA_COMPACTED,
-						fmt.Sprintf("required watch revision: %d is smaller than current compact/min revision %d.", startRevision, res.CompactRevision))
-					resp.CompactRevision = res.CompactRevision
-				} else {
-					resp.Header = s.wrapErrorToHeader(pdpb.ErrorType_UNKNOWN,
-						fmt.Sprintf("watch channel meet other error %s.", res.Err().Error()))
-				}
-				if err := server.Send(&resp); err != nil {
-					return err
-				}
-				// Err() indicates that this WatchResponse holds a channel-closing error.
-				return res.Err()
-			}
-
-			events := make([]*pdpb.Event, 0, len(res.Events))
-			for _, e := range res.Events {
-				event := &pdpb.Event{Kv: &pdpb.KeyValue{Key: e.Kv.Key, Value: e.Kv.Value}, Type: pdpb.Event_EventType(e.Type)}
-				if e.PrevKv != nil {
-					event.PrevKv = &pdpb.KeyValue{Key: e.PrevKv.Key, Value: e.PrevKv.Value}
-				}
-				events = append(events, event)
-			}
-			if len(events) > 0 {
-				if err := server.Send(&pdpb.WatchResponse{Events: events, CompactRevision: res.Header.GetRevision()}); err != nil {
-					return err
-				}
-			}
-		}
-	}
-}
-
-// Get gets the key-value pair with a given key.
-func (s *GrpcServer) Get(ctx context.Context, req *pdpb.GetRequest) (*pdpb.GetResponse, error) {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-	options := []clientv3.OpOption{}
-	key := string(req.GetKey())
-	if endKey := req.GetRangeEnd(); endKey != nil {
-		options = append(options, clientv3.WithRange(string(endKey)))
-	}
-	if rev := req.GetRevision(); rev != 0 {
-		options = append(options, clientv3.WithRev(rev))
-	}
-	if limit := req.GetLimit(); limit != 0 {
-		options = append(options, clientv3.WithLimit(limit))
-	}
-	res, err := s.client.Get(ctx, key, options...)
-	if err != nil {
-		return &pdpb.GetResponse{Header: s.wrapErrorToHeader(pdpb.ErrorType_UNKNOWN, err.Error())}, nil
-	}
-	resp := &pdpb.GetResponse{
-		Header: s.header(),
-		Count:  res.Count,
-		More:   res.More,
-	}
-	if res.Header != nil {
-		resp.Revision = res.Header.GetRevision()
-	}
-	for _, kv := range res.Kvs {
-		resp.Kvs = append(resp.Kvs, &pdpb.KeyValue{Key: kv.Key, Value: kv.Value})
-	}
-
-	return resp, nil
-}
-
-// Put puts the key-value pair into etcd.
-func (s *GrpcServer) Put(ctx context.Context, req *pdpb.PutRequest) (*pdpb.PutResponse, error) {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-	options := []clientv3.OpOption{}
-	key := string(req.GetKey())
-	value := string(req.GetValue())
-	if lease := clientv3.LeaseID(req.GetLease()); lease != 0 {
-		options = append(options, clientv3.WithLease(lease))
-	}
-	if prevKv := req.GetPrevKv(); prevKv {
-		options = append(options, clientv3.WithPrevKV())
-	}
-
-	res, err := s.client.Put(ctx, key, value, options...)
-	if err != nil {
-		return &pdpb.PutResponse{Header: s.wrapErrorToHeader(pdpb.ErrorType_UNKNOWN, err.Error())}, nil
-	}
-
-	resp := &pdpb.PutResponse{
-		Header: s.header(),
-	}
-	if res.Header != nil {
-		resp.Revision = res.Header.GetRevision()
-	}
-	if res.PrevKv != nil {
-		resp.PrevKv = &pdpb.KeyValue{Key: res.PrevKv.Key, Value: res.PrevKv.Value}
-	}
-	return resp, nil
-}
-
 // Evict the leaders when the store is damaged. Damaged regions are emergency errors
 // and requires user to manually remove the `evict-leader-scheduler` with pd-ctl
 func (s *GrpcServer) handleDamagedStore(stats *pdpb.StoreStats) {
