+++ tests/client/go.mod
@@ -172,4 +172,4 @@ replace (
 // reset grpc and protobuf deps in order to import client and server at the same time
 replace google.golang.org/grpc v1.51.0 => google.golang.org/grpc v1.26.0
 
-replace github.com/pingcap/kvproto => github.com/rleungx/kvproto v0.0.0-20230223090544-2d5287a102d2
+replace github.com/pingcap/kvproto => github.com/rleungx/kvproto v0.0.0-20230224062044-754da3291461
