+++ tools/pd-tso-bench/go.mod
@@ -13,4 +13,4 @@ require (
 
 replace github.com/tikv/pd/client => ../../client
 
-replace github.com/pingcap/kvproto => github.com/rleungx/kvproto v0.0.0-20230223090544-2d5287a102d2
+replace github.com/pingcap/kvproto => github.com/rleungx/kvproto v0.0.0-20230224062044-754da3291461
